scap inlining refactoring, the entry point is still not inlined, wip
